refactor: change name of folder containing majority of core script code
