+++ .github/cicd/core/IDotNetTool.ts
