+++ .github/cicd/core/VersionSidebarService.ts
